weviate/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
apturl==0.5.2
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brlapi==0.8.3
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.5.14
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
evdev==1.6.1
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fasteners==0.14.1
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.2
google-search-results==2.4.2
greenlet==2.0.2
h11==0.14.0
httplib2==0.20.2
huggingface-hub==0.16.4
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
keyring==23.5.0
langchain==0.0.266
langchain-experimental==0.0.10
langsmith==0.0.24
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
openai==0.27.8
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.12
PyGObject==3.42.1
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==2.4.7
pypdf==3.15.2
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-xlib==0.33
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2023.8.8
reportlab==3.6.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
SecretStorage==3.3.1
selenium==4.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sympy==1.12
systemd-python==234
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.32.0
trio==0.22.2
trio-websocket==0.10.3
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
validators==0.21.0
wadllib==1.3.6
weaviate-client==3.23.1
Werkzeug==2.3.6
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
